requirements.txt
django>=1.8.18
# Additional requirements go here
python-telegram-bot>=6.0.1
